use size_t rather than intptr_t
after some discussion keeping to a more standard type was decided as a good idea, size_t is probably what we should have used from the get go thinking about it

+++ Hazel/src/Hazel/Renderer/Buffer.h
@@ -33,7 +33,7 @@ namespace Hazel {
 		std::string Name;
 		ShaderDataType Type;
 		uint32_t Size;
-		intptr_t Offset;
+		size_t Offset;
 		bool Normalized;
 
 		BufferElement() = default;
@@ -86,7 +86,7 @@ namespace Hazel {
 	private:
 		void CalculateOffsetsAndStride()
 		{
-			uint32_t offset = 0;
+			size_t offset = 0;
 			m_Stride = 0;
 			for (auto& element : m_Elements)
 			{
